+++ Changelog.md
@@ -2,3 +2,4 @@
 #### Version 1.0.0-Beta.2
 + Features
   + Folder Selection Buttons added
+  + Initial Setup for Eth2 beacon chain and validator added